requirements.txt
pillow==10.0.0
opencv-python==4.8.0.74
python-dotenv==1.0.0
bcrypt==4.0.1